chore(update): update text (unsure, please review this!)
We need to use `TextureAtlasLayout` instead of `TextureAtlas` now due to
a refactor of this. This works mostly fine, but in the end we're missing
the `Handle<Image>` which was previously given with the `TextureAtlas`.
I changed that to the image handle of a random glyph assuming they all
share a texture (the original `TextureAtlas`'s texture).

I'm not sure though.

Authored-by: RobWalt <robwalter96@gmail.com>